chore(generator): Remove use of installComplete
The Falcon Yeoman Generator class had sort of been using the member
variable "installComplete" to determine whether a command was "fully"
successful or if it completed with warnings. We have removed it from use
and switched to inspecting whether the GeneratorStatus object has an
warnings once we reach the end() or _default_end() functions.